+++ test_scripts/Polices/Validation_of_PolicyTables/242_ATF_merge_preloaded_pt_into_local_pt_device_data.lua
@@ -1,0 +1,333 @@
+---------------------------------------------------------------------------------------------
+-- Requirement summary:
+-- [Policies] Merging rules for "device_data" section
+--
+-- Description:
+-- Check of merging rules for "device_data" section
+-- 1. Used preconditions
+-- Delete files and policy table from previous ignition cycle if any
+-- Start SDL with PreloadedPT json file with "preloaded_date" parameter
+-- Change data in "device_data" section in LocalPT
+-- 2. Performed steps
+-- Stop SDL
+-- Start SDL with corrected PreloadedPT json file with "preloaded_date" parameter with bigger value
+--
+-- Expected result:
+-- SDL must leave all fields & their values of "device_data" section as it was in the database without changes
+---------------------------------------------------------------------------------------------
+
+--[[ General configuration parameters ]]
+config.deviceMAC = "12ca17b49af2289436f303e0166030a21e525d266e209267433801a8fd4071a0"
+
+--[[ Required Shared libraries ]]
+local commonFunctions = require ('user_modules/shared_testcases/commonFunctions')
+local commonSteps = require ('user_modules/shared_testcases/commonSteps')
+local commonPreconditions = require ('user_modules/shared_testcases/commonPreconditions')
+local testCasesForPolicyTable = require ('user_modules/shared_testcases/testCasesForPolicyTable')
+local json = require("modules/json")
+
+--[[ General configuration parameters ]]
+Test = require('connecttest')
+local config = require('config')
+require('user_modules/AppTypes')
+config.defaultProtocolVersion = 2
+config.application1.registerAppInterfaceParams.deviceInfo.hardware = "Nexus"
+
+--[[ Local Variables ]]
+local Device_timestamp
+local TESTED_DATA = {
+  preloaded_date = {"1988-12-01","2015-05-02"},
+  device = {
+    id = tostring(config.deviceMAC),
+    hardware = tostring(config.application1.registerAppInterfaceParams.deviceInfo.hardware),
+    firmware_rev = tostring(config.application1.registerAppInterfaceParams.deviceInfo.firmwareRev),
+    os = tostring(config.application1.registerAppInterfaceParams.deviceInfo.os),
+    os_version = tostring(config.application1.registerAppInterfaceParams.deviceInfo.osVersion),
+    carrier = tostring(config.application1.registerAppInterfaceParams.deviceInfo.carrier),
+    max_number_rfcom_ports = tostring(config.application1.registerAppInterfaceParams.deviceInfo.maxNumberRFCOMMPorts),
+    user_consent_records = "not_available"
+  },
+  device_consent_group = {
+    device_id = tostring(config.deviceMAC),
+    functional_group_id = "DataConsent-2",
+    is_consented = "1",
+    input = "GUI",
+    time_stamp = ""
+  }
+}
+local PRELOADED_PT_FILE_NAME = "sdl_preloaded_pt.json"
+
+local TestData = {
+  path = config.pathToSDL .. "TestData",
+  isExist = false,
+  init = function(self)
+    if not self.isExist then
+      os.execute("mkdir ".. self.path)
+      os.execute("echo 'List test data files files:' > " .. self.path .. "/index.txt")
+      self.isExist = true
+    end
+  end,
+  store = function(self, message, pathToFile, fileName)
+    if self.isExist then
+      local dataToWrite = message
+
+      if pathToFile and fileName then
+        os.execute(table.concat({"cp ", pathToFile, " ", self.path, "/", fileName}))
+        dataToWrite = table.concat({dataToWrite, " File: ", fileName})
+      end
+
+      dataToWrite = dataToWrite .. "\n"
+      local file = io.open(self.path .. "/index.txt", "a+")
+      file:write(dataToWrite)
+      file:close()
+    end
+  end,
+  delete = function(self)
+    if self.isExist then
+      os.execute("rm -r -f " .. self.path)
+      self.isExist = false
+    end
+  end,
+  info = function(self)
+    if self.isExist then
+      commonFunctions:userPrint(35, "All test data generated by this test were stored to folder: " .. self.path)
+    else
+      commonFunctions:userPrint(35, "No test data were stored" )
+    end
+  end
+}
+
+--[[ Local Functions ]]
+local function updatePreloadedPt(updaters)
+  local pathToFile = config.pathToSDL .. PRELOADED_PT_FILE_NAME
+  local file = io.open(pathToFile, "r")
+  local json_data = file:read("*a")
+  file:close()
+
+  local data = json.decode(json_data)
+  if data then
+    for _, updateFunc in pairs(updaters) do
+      updateFunc(data)
+    end
+  end
+
+  local dataToWrite = json.encode(data)
+  file = io.open(pathToFile, "w")
+  file:write(dataToWrite)
+  file:close()
+end
+
+local function prepareInitialPreloadedPT()
+  local initialUpdaters = {
+    function(data)
+      for key, value in pairs(data.policy_table.functional_groupings) do
+        if not value.rpcs then
+          data.policy_table.functional_groupings[key] = nil
+        end
+      end
+    end,
+    function(data)
+      data.policy_table.module_config.preloaded_date = TESTED_DATA.preloaded_date[1]
+    end
+  }
+  updatePreloadedPt(initialUpdaters)
+end
+
+local function prepareNewPreloadedPT()
+  local newUpdaters = {
+    function(data)
+      data.policy_table.module_config.preloaded_date = TESTED_DATA.preloaded_date[2]
+    end,
+  }
+  updatePreloadedPt(newUpdaters)
+end
+
+local function constructPathToDatabase()
+  if commonSteps:file_exists(config.pathToSDL .. "storage/policy.sqlite") then
+    return config.pathToSDL .. "storage/policy.sqlite"
+  elseif commonSteps:file_exists(config.pathToSDL .. "policy.sqlite") then
+    return config.pathToSDL .. "policy.sqlite"
+  else
+    commonFunctions:userPrint(31, "policy.sqlite is not found" )
+    return nil
+  end
+end
+
+local function executeSqliteQuery(rawQueryString, dbFilePath)
+  if not dbFilePath then
+    return nil
+  end
+  local queryExecutionResult = {}
+  local queryString = table.concat({"sqlite3 ", dbFilePath, " '", rawQueryString, "'"})
+  local file = io.popen(queryString, 'r')
+  if file then
+    local index = 1
+    for line in file:lines() do
+      queryExecutionResult[index] = line
+      index = index + 1
+    end
+    file:close()
+    return queryExecutionResult
+  else
+    return nil
+  end
+end
+
+local function isValuesCorrect(actualValues, expectedValues)
+  if #actualValues ~= #expectedValues then
+    return false
+  end
+
+  local tmpExpectedValues = {}
+  for i = 1, #expectedValues do
+    tmpExpectedValues[i] = expectedValues[i]
+  end
+
+  local isFound
+  for j = 1, #actualValues do
+    isFound = false
+    for key, value in pairs(tmpExpectedValues) do
+      if value == actualValues[j] then
+        isFound = true
+        tmpExpectedValues[key] = nil
+        break
+      end
+    end
+    if not isFound then
+      return false
+    end
+  end
+  if next(tmpExpectedValues) then
+    return false
+  end
+  return true
+end
+
+function Test.checkLocalPT(checkTable)
+  local expectedLocalPtValues
+  local queryString
+  local actualLocalPtValues
+  local comparationResult
+  local isTestPass = true
+  for _, check in pairs(checkTable) do
+    expectedLocalPtValues = check.expectedValues
+    queryString = check.query
+    actualLocalPtValues = executeSqliteQuery(queryString, constructPathToDatabase())
+    if actualLocalPtValues then
+      comparationResult = isValuesCorrect(actualLocalPtValues, expectedLocalPtValues)
+      if not comparationResult then
+        TestData:store(table.concat({"Test ", queryString, " failed: SDL has wrong values in LocalPT"}))
+        TestData:store("ExpectedLocalPtValues")
+        commonFunctions:userPrint(31, table.concat({"Test ", queryString, " failed: SDL has wrong values in LocalPT"}))
+        commonFunctions:userPrint(35, "ExpectedLocalPtValues")
+        for _, values in pairs(expectedLocalPtValues) do
+          TestData:store(values)
+          print(values)
+        end
+        TestData:store("ActualLocalPtValues")
+        commonFunctions:userPrint(35, "ActualLocalPtValues")
+        for _, values in pairs(actualLocalPtValues) do
+          TestData:store(values)
+          print(values)
+        end
+        isTestPass = false
+      end
+    else
+      TestData:store("Test failed: Can't get data from LocalPT")
+      commonFunctions:userPrint(31, "Test failed: Can't get data from LocalPT")
+      isTestPass = false
+    end
+  end
+  return isTestPass
+end
+
+--[[ General Precondition before ATF start ]]
+config.defaultProtocolVersion = 2
+testCasesForPolicyTable.Delete_Policy_table_snapshot()
+commonSteps:DeleteLogsFileAndPolicyTable()
+--testCasesForPolicyTable:Precondition_updatePolicy_By_overwriting_preloaded_pt("files/jsons/Policies/PTU_ValidationRules/pre_dataconsent.json")
+commonPreconditions:BackupFile(PRELOADED_PT_FILE_NAME)
+prepareInitialPreloadedPT()
+
+--[[ Preconditions ]]
+commonFunctions:newTestCasesGroup("Precondition")
+function Test:Precondition_trigger_getting_device_consent()
+  testCasesForPolicyTable:trigger_getting_device_consent(self, config.application1.registerAppInterfaceParams.appName, config.deviceMAC)
+end
+
+--[[ Test ]]
+commonFunctions:newTestCasesGroup("Test")
+
+function Test:TestStep_VerifyInitialLocalPT()
+  os.execute("sleep 3")
+  TestData:store("Initial Local PT is stored", constructPathToDatabase(), "initial_policy.sqlite")
+  local checks = {
+    { query = 'select preloaded_date from module_config', expectedValues = {TESTED_DATA.preloaded_date[1]} },
+    { query = 'select id from device', expectedValues = {TESTED_DATA.device.id} },
+    { query = 'select hardware from device', expectedValues = {TESTED_DATA.device.hardware} },
+    { query = 'select firmware_rev from device', expectedValues = {TESTED_DATA.device.firmware_rev} },
+    { query = 'select os from device', expectedValues = {TESTED_DATA.device.os} },
+    { query = 'select os_version from device', expectedValues = {TESTED_DATA.device.os_version} },
+    { query = 'select carrier from device', expectedValues = {TESTED_DATA.device.carrier} },
+    { query = 'select max_number_rfcom_ports from device', expectedValues = {TESTED_DATA.device.max_number_rfcom_ports} },
+    { query = 'select user_consent_records from device', expectedValues = {TESTED_DATA.device.user_consent_records} },
+    { query = 'select device_id from device_consent_group', expectedValues = {TESTED_DATA.device_consent_group.device_id} },
+    { query = 'select functional_group_id from device_consent_group', expectedValues = {TESTED_DATA.device_consent_group.functional_group_id} },
+    { query = 'select is_consented from device_consent_group', expectedValues = {TESTED_DATA.device_consent_group.is_consented} },
+    { query = 'select input from device_consent_group', expectedValues = {TESTED_DATA.device_consent_group.input} }
+  }
+
+  local Device_table = executeSqliteQuery('select time_stamp from device_consent_group', constructPathToDatabase())
+  for _, v in pairs(Device_table) do
+    Device_timestamp = v
+  end 
+  
+  if not self.checkLocalPT(checks) then
+    self:FailTestCase("SDL has wrong values in LocalPT")
+  end
+end
+
+function Test:TestStep_StopSDL()
+  StopSDL(self)
+end
+
+function Test.TestStep_LoadNewPreloadedPT()
+  prepareNewPreloadedPT()
+end
+
+function Test:TestStep_StartSDL()
+  StartSDL(config.pathToSDL, true, self)
+end
+
+function Test:TestStep_VerifyNewLocalPT()
+  os.execute("sleep 3")
+  TestData:store("New Local PT is stored", constructPathToDatabase(), "new_policy.sqlite")
+  local checks = {
+    { query = 'select preloaded_date from module_config', expectedValues = {TESTED_DATA.preloaded_date[2]}},
+    { query = 'select id from device', expectedValues = {TESTED_DATA.device.id} },
+    { query = 'select hardware from device', expectedValues = {TESTED_DATA.device.hardware} },
+    { query = 'select firmware_rev from device', expectedValues = {TESTED_DATA.device.firmware_rev} },
+    { query = 'select os from device', expectedValues = {TESTED_DATA.device.os} },
+    { query = 'select os_version from device', expectedValues = {TESTED_DATA.device.os_version} },
+    { query = 'select carrier from device', expectedValues = {TESTED_DATA.device.carrier} },
+    { query = 'select max_number_rfcom_ports from device', expectedValues = {TESTED_DATA.device.max_number_rfcom_ports} },
+    { query = 'select user_consent_records from device', expectedValues = {TESTED_DATA.device.user_consent_records} },
+    { query = 'select device_id from device_consent_group', expectedValues = {TESTED_DATA.device_consent_group.device_id} },
+    { query = 'select functional_group_id from device_consent_group', expectedValues = {TESTED_DATA.device_consent_group.functional_group_id} },
+    { query = 'select is_consented from device_consent_group', expectedValues = {TESTED_DATA.device_consent_group.is_consented} },
+    { query = 'select input from device_consent_group', expectedValues = {TESTED_DATA.device_consent_group.input} },
+    { query = 'select time_stamp from device_consent_group', expectedValues = {Device_timestamp} },
+  }
+  if not self.checkLocalPT(checks) then
+    self:FailTestCase("SDL has wrong values in LocalPT")
+  end
+end
+
+--[[ Postconditions ]]
+commonFunctions:newTestCasesGroup("Postconditions")
+testCasesForPolicyTable:Restore_preloaded_pt()
+function Test.Postcondition()
+  StopSDL()
+end
+
+return Test
